+++ ftmq/util.py
@@ -1,8 +1,10 @@
-from typing import Generator
+from collections.abc import Generator
+from functools import cache
 
 from nomenklatura.dataset import DataCatalog, Dataset
 
 
+@cache
 def make_dataset(name: str) -> Dataset:
     catalog = DataCatalog(
         Dataset, {"datasets": [{"name": name, "title": name.title()}]}