+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## CHANGELOG
 
+### 0.2.1 (2023-08-07)
+
+- Align with `nomenklatura` v3
+
 ### 0.2.0 (2023-07-31)
 
-- Model [nomenklatura](github.com/opensanctions/nomenklatura) catalog/dataset via pydantic
+- Model [nomenklatura](https://github.com/opensanctions/nomenklatura) catalog/dataset via pydantic
 - Replace `smart_open` with [fsspec](https://github.com/fsspec/filesystem_spec)
 - add generic IO handling based on `fsspec`